Add files via upload
This project will tell how much a student can score according to the hours he studies.
Linear Regression used, In this project by scatter graph clearly show that score is directly proportional to the study of the hour. Training and testing data is taken in a 1:4 ratio. The accuracy was of this project 78.48. the intercept of an hour and the study was 1.98 and the coefficient was 9.82. The shape of the data was [25*2].